refactor(web): extract PairedShellPane from RightPanel
Move the paired (side-shell) xterm.js terminal and its host/container
switch out of RightPanel into PairedTerminal.tsx as a reusable
PairedShellPane, with a fullViewport prop that selects the keyboard
padding posture. No behavior change yet; the mobile picker work in #1452
reuses this as a standalone full-viewport pane.
